Play 2.4: can now add comments through Validate or LabelMap

+++ app/formats/json/CommentSubmissionFormats.scala
@@ -10,11 +10,11 @@ object CommentSubmissionFormats {
 
   case class ValidationCommentSubmission(missionId: Int, labelId: Int, comment: String,
                                          gsvPanoramaId: String, heading: Double, pitch: Double,
-                                         zoom: Int, lat: Double, lng: Double)
+                                         zoom: Float, lat: Double, lng: Double)
 
-  case class LabelMapValidationCommentSubmission(labelId: Int, labelType: String, comment: String, 
-                                        gsvPanoramaId: String, heading: Double, pitch: Double, zoom: Int, 
-                                        lat: Double, lng: Double)
+  case class LabelMapValidationCommentSubmission(labelId: Int, labelType: String, comment: String,
+                                                 gsvPanoramaId: String, heading: Double, pitch: Double, zoom: Float,
+                                                 lat: Double, lng: Double)
 
   implicit val commentSubmissionReads: Reads[CommentSubmission] = (
     (JsPath \ "audit_task_id").read[Int] and
@@ -36,7 +36,7 @@ object CommentSubmissionFormats {
       (JsPath \ "gsv_panorama_id").read[String] and
       (JsPath \ "heading").read[Double] and
       (JsPath \ "pitch").read[Double] and
-      (JsPath \ "zoom").read[Int] and
+      (JsPath \ "zoom").read[Float] and
       (JsPath \ "lat").read[Double] and
       (JsPath \ "lng").read[Double]
   )(ValidationCommentSubmission.apply _)
@@ -48,7 +48,7 @@ object CommentSubmissionFormats {
       (JsPath \ "gsv_panorama_id").read[String] and
       (JsPath \ "heading").read[Double] and
       (JsPath \ "pitch").read[Double] and
-      (JsPath \ "zoom").read[Int] and
+      (JsPath \ "zoom").read[Float] and
       (JsPath \ "lat").read[Double] and
       (JsPath \ "lng").read[Double]
   )(LabelMapValidationCommentSubmission.apply _)
